% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colwell_stats.R
\name{colwell_stats}
\alias{colwell_stats}
\title{Environmental time series metrics from Colwell (1974)}
\usage{
colwell_stats(time_series, dates, n_states = 11)
}
\arguments{
\item{time_series}{A \code{\link[base]{numeric}} vector containing
a raw environmental time series.}

\item{dates}{An vector of class \code{\link[base]{Date}}
of format YYYY-MM-DD (must be in progressive chronological order).}

\item{n_states}{is a \code{\link[base]{numeric}} vector of length 1 containing
a somewhat arbitrary number, as Colwell's method divides a continuous variable
up into discrete states. Default (arbitrary) is 11. See Details.}
}
\value{
A \code{\link[base]{data.frame}}.
}
\description{
Calculates Constancy (C), Contingency (M), and Predictability (P) metrics from Colwell (1974)
for seasonally varying environmental data.
}
\examples{
library(envPred)
data(sst)
data(npp)
n_states <- 11
fct <- envPred:::colwell_stats
fct(sst$time_series, sst$dates, n_states)
fct(npp$time_series, npp$dates, n_states)
}
\seealso{
\code{\link{seasonality_and_colour}}, \code{\link{env_stats}}.
}
\author{
Diego Barneche and Scott Burgess.
}
