% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{detrend_and_bin}
\alias{detrend_and_bin}
\title{Binds output from \code{\link{linear_detrend}} and
\code{\link{monthly_bins}}}
\usage{
detrend_and_bin(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{linear_detrend}}}
}
\value{
A \code{\link[base]{data.frame}}.
}
\description{
Wrapper
}
\author{
Diego Barneche and Scott Burgess.
}
