% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{envNoise}
\alias{envNoise}
\title{Calculate the 'color' of the noise of unpredicted detrended residuals}
\usage{
envNoise(residualTimeSeries, predictor, checkPlots,
  noiseMethod = c("spectrum", "LombScargle"))
}
\arguments{
\item{residualTimeSeries}{A \code{\link[base]{numeric}} vector containing unpredicted 
detrended residuals as obtained by function \code{\link[envPred]{monthlyBinning}}.}

\item{predictor}{A \code{\link[base]{numeric}} vector containing the time (in days)
for each observation in \code{residualTimeSeries}. Needs to start at day 0.}

\item{checkPlots}{Should inspection plots be triggered? No default value.}

\item{noiseMethod}{A method for estimating the slope beta. Takes 2 possible 
values: \code{'spectrum'} for evenly distributed time series or 
\code{'LombScargle'} for unevenly distributed ones.}
}
\value{
A \code{\link[base]{numeric}} vector of length 1 containing the beta slope.
}
\description{
Calculate environmental noise
}
\details{
This function calculates beta slope (assuming 1/f noise family) on the
residual time series using function \code{\link[envPred]{extractSpectrumSlope}}.
}
\seealso{
\code{\link[envPred]{monthlyBinning}}, \code{\link[envPred]{envPredictability}}, \code{\link[envPred]{extractSpectrumSlope}}.
}
\author{
Diego Barneche and Scott Burgess.
}
