% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envpreddata-class.R
\docType{class}
\name{envpreddata-class}
\alias{envpreddata-class}
\alias{envpreddata}
\title{Class \code{envpreddata} of environmental time series statistics calculated with the \pkg{envPred} package}
\description{
Statistics calculated with the \code{\link[envPred:envPred]{envPred}} package are
represented as a \code{envpreddata} object, which contains a data.frame with
descriptive summary of the time series vector (length, number of NAs, etc), the
mean, variance, and coefficient of variation of the environmental time series, and
seasonality, colour of environmental noise, constancy, contingency and predictability.
}
\details{
See function \code{\link{env_stats}}.
}
\section{Slots}{

\describe{
\item{\code{series_n}}{Length of the time series.}

\item{\code{n_na}}{Number of NAs in the time series.}

\item{\code{prop_na}}{Proportion of NAs in the time series.}

\item{\code{n_yrs}}{Number of years in the time series.}

\item{\code{n_months}}{Number of months in the time series.}

\item{\code{n_days}}{Number of days in the time series.}

\item{\code{frequency}}{Frequency (\eqn{2 / (n \Delta t)}) of the time series, where \eqn{\Delta t} is the time gap between consecutive points in the time series, and \eqn{n} is the number of observations in the time series.}

\item{\code{nyquist_freq}}{Nyquist frequency (\eqn{1 / (2 \Delta t)}) of the time series.}

\item{\code{raw_mean}}{Arithmetic mean of the time series.}

\item{\code{raw_var}}{Variance of the time series.}

\item{\code{raw_cv}}{Coefficient of variation (standard deviation / mean) of the time series.}

\item{\code{predicted_var}}{Variance of the seasonal trend. See \code{\link{env_stats}} for further explanations.}

\item{\code{unpredicted_var}}{Variance of the residual time series (i.e. the time series after the seasonal trend was removed). See \code{\link{env_stats}} for further explanations.}

\item{\code{unbounded_seasonality}}{Seasonality as var_predict / var_unpredict. See \code{\link{env_stats}} for further explanations.}

\item{\code{bounded_seasonality}}{Seasonality as var_predict / (var_predict + var_unpredict). See \code{\link{env_stats}} for further explanations.}

\item{\code{env_col}}{Colour of environmental noise. See \code{\link{env_stats}} for further explanations.}

\item{\code{colwell_c}}{Colwell's constancy. See \code{\link{env_stats}} for further explanations.}

\item{\code{colwell_m}}{Colwell's contingency. See \code{\link{env_stats}} for further explanations.}

\item{\code{colwell_p}}{Colwell's predictability. See \code{\link{env_stats}} for further explanations.}
}}

\seealso{
\code{\link{env_stats}},
\code{\link{gg_envpred}}
}
