% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{envPredictability}
\alias{envPredictability}
\title{Calculates environmental predictability components

Environmental predictability components (deprecated).
Use \code{\link{predictability}} instead.}
\usage{
envPredictability(
  rawTimeSeries,
  datesVector,
  delta,
  isUneven = FALSE,
  interpolate = FALSE,
  checkPlots = FALSE,
  showWarnings = TRUE,
  noiseMethod
)
}
\arguments{
\item{rawTimeSeries}{A \code{\link[base]{numeric}} vector containing
a raw environmental time series.}

\item{datesVector}{An vector of class \code{\link[base]{Date}} of format YYYY-MM-DD
(must be in progressive chronological order).}

\item{delta}{Time interval (any unit) of \code{rawTimeSeries}.}

\item{isUneven}{Is \code{rawTimeSeries} even or unevenly distributed in time? Default is \code{\link[base]{FALSE}}.}

\item{interpolate}{Should a linear interpolation be applied to missing values in
\code{rawTimeSeries}? Irrelevant method if \code{isUneven} is \code{\link[base]{FALSE}}. Default is \code{\link[base]{FALSE}}.}

\item{checkPlots}{Deprecated. Use newer \code{\link{gg_envpred}} instead.}

\item{showWarnings}{Should cautionary warning messages be displayed? Default is \code{\link[base]{TRUE}}.
Strongly recommended for first time users.}

\item{noiseMethod}{A method for estimating the slope beta. Takes 2 possible
values: \code{'spectrum'} for evenly distributed time series or
\code{'LombScargle'} for unevenly distributed ones.}
}
\value{
A \code{\link[base]{data.frame}} with environmental predictability components.
}
\description{
Calculates environmental predictability components

Environmental predictability components (deprecated).
Use \code{\link{predictability}} instead.
}
\seealso{
\code{\link{predictability}}.
}
\author{
Diego Barneche and Scott Burgess.
}
