% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{envSeasonality}
\alias{envSeasonality}
\title{Calculate the 'color' of unpredicted detrended residuals}
\usage{
envSeasonality(interpolatedSeasons, residualTimeSeries,
  seasonalityMethod = c("absolute", "unbounded", "bounded"))
}
\arguments{
\item{interpolatedSeasons}{A \code{\link[base]{numeric}} vector containing predicted (interpolated) 
detrended residuals as obtained by function \code{\link[envPred]{monthlyBinning}}.}

\item{residualTimeSeries}{A \code{\link[base]{numeric}} vector containing unpredicted 
detrended residuals as obtained by function \code{\link[envPred]{monthlyBinning}}.}

\item{seasonalityMethod}{A method for estimating seasonality. Takes 3 possible 
values: \code{'absolute'}, \code{'unbounded'}, or \code{'bounded'}, with no default.}
}
\value{
A \code{\link[base]{list}} containing the sample variance of \code{interpolatedSeasons},
\code{residualTimeSeries}, and the resulting seasonality. See details.
}
\description{
Calculate environmental seasonality
}
\details{
If \code{seasonalityMethod} is \code{'absolute'}, then seasonality corresponds to the
sample variance of \code{interpolatedSeasons}; if \code{'unbounded'}, it corresponds to to the 
ratio between sample variances of \code{interpolatedSeasons} and \code{residualTimeSeries};
if \code{'bounded'}, it corresponds to the sample variance of \code{interpolatedSeasons}
relative to the total summed variances of both \code{interpolatedSeasons} and \code{residualTimeSeries}.
}
\seealso{
\code{\link[envPred]{monthlyBinning}}, \code{\link[envPred]{envPredictability}}.
}
\author{
Diego Barneche and Scott Burgess.
}
