% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_envpred.R
\name{gg_envpred}
\alias{gg_envpred}
\title{Plot an object of class \code{envpreddata}}
\usage{
gg_envpred(object, type = c("detrended", "spectral"))
}
\arguments{
\item{object}{An object of class \code{\link{envpreddata}}}

\item{type}{Either \code{detrended} (default) or \code{spectral}.
See Details.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot an object of class \code{envpreddata}
}
\details{
Two types of plots are provided:
detrended: show the time series after linear
de-trending, including the seasonality interpolation.
spectral: The spectral power density as
a function of frequency.
}
\examples{
\dontrun{
library(envPred)
data(sst)
dat <- env_stats(sst$time_series,
                 sst$dates,
                 n_states = 11,
                 delta = 1,
                 is_uneven = FALSE,
                 interpolate = FALSE,
                 show_warns = TRUE,
                 noise_method = 'spectrum')

gg_envpred(dat, type = "detrended")
gg_envpred(dat, type = "spectral")
}
}
\seealso{
\code{\link{env_stats}}, \code{\link{envpreddata}}.
}
