% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is.envpreddata}
\alias{is.envpreddata}
\title{Checks if argument is an \code{envpreddata} object}
\usage{
is.envpreddata(x)
}
\arguments{
\item{x}{An R object}
}
\description{
Checks if argument is an \code{envpreddata} object
}
\seealso{
\code{\link{envpreddata}}, \code{\link{env_stats}}.
}
