% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_envpreddata}
\alias{is_envpreddata}
\title{Checks if argument is an \code{envpreddata} object}
\usage{
is_envpreddata(x)
}
\arguments{
\item{x}{An R object}
}
\description{
Checks if argument is an \code{envpreddata} object
}
