% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_predictability}
\alias{is_predictability}
\title{Checks if argument is a \code{predictability} object}
\usage{
is_predictability(x)
}
\arguments{
\item{x}{An \R object}
}
\description{
Checks if argument is a \code{predictability} object
}
