% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{last_day}
\alias{last_day}
\title{Compute last day of the month}
\usage{
last_day(x)
}
\arguments{
\item{x}{An object of class \code{\link[base]{Date}} of format YYYY-MM-DD
(must be in progressive chronological order).}
}
\value{
An object of class \code{\link[base]{Date}} of format YYYY-MM-DD.
}
\description{
Compute last day of the month
}
\seealso{
\code{\link{monthly_bins}}.
}
\author{
Diego Barneche and Scott Burgess.
}
