% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{linear_detrend}
\alias{linear_detrend}
\title{Linear detrending of time series}
\usage{
linear_detrend(time_series, dates)
}
\arguments{
\item{time_series}{A \code{\link[base]{numeric}} vector containing
a raw environmental time series.}

\item{dates}{An object of class \code{\link[base]{Date}} of format YYYY-MM-DD
(must be in progressive chronological order).}
}
\value{
A \code{\link[base]{data.frame}} containing a \code{\link[base]{numeric}} vector
of the time (in days) for each observation in \code{time_series} starting at day 0,
and a \code{\link[base]{numeric}} vector containing the residual variation of \code{time_series}
after removing the linear trend.
}
\description{
Linear detrending of time series
}
\seealso{
\code{\link{seasonality_and_colour}}.
}
\author{
Diego Barneche and Scott Burgess.
}
