% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lmb_output_prep}
\alias{lmb_output_prep}
\title{Adapt Lomb-Scargle Periodogram}
\usage{
lmb_output_prep(resid_time_series, ...)
}
\arguments{
\item{resid_time_series}{A \code{\link[base]{numeric}} vector containing unpredicted
de-trended residuals as obtained by function \code{\link{monthly_bins}}.}

\item{...}{Additional arguments to \code{\link[lomb]{lsp}}.}
}
\value{
A \code{\link[base]{data.frame}} containing a vector
of frequencies/periods scanned, and a vector containing the normalised power
corresponding to scanned frequencies/periods.
}
\description{
Calculates and renames output of Lomb-Scargle Periodogram
}
\seealso{
\code{\link{spec_slope_get}}, \code{\link{seasonality_and_colour}}.
}
\author{
Diego Barneche and Scott Burgess.
}
