% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{monthlyBinning}
\alias{monthlyBinning}
\title{Compute the seasonal (monthly) components of time series}
\usage{
monthlyBinning(resids, datesVec)
}
\arguments{
\item{resids}{A \code{\link[base]{numeric}} vector containing the residual
variation of a raw time series after removing the linear trend.}

\item{datesVec}{An object of class \code{\link[base]{Date}} of format YYYY-MM-DD
(must be in progressive chronological order).}
}
\value{
A \code{\link[base]{list}} containing the a \code{\link[base]{numeric}}
vector containing predicted (interpolated) detrended residuals (\code{'interpolatedSeasons'}), and
a \code{\link[base]{numeric}} vector containing unpredicted detrended residuals
\code{'residualTimeSeries'}. See details.
}
\description{
Compute the seasonal (monthly) components of time series
}
\details{
This algorithm follows the steps described in \href{http://onlinelibrary.wiley.com/doi/10.1111/ele.12402/abstract}{Marshall and Burgess 2015} Ecology Letters 18: 1461–0248, doi: \href{http://dx.doi.org/10.1111/ele.12402}{10.1111/ele.12402}.
}
\seealso{
\code{\link[envPred]{envPredictability}}.
}
\author{
Diego Barneche and Scott Burgess.
}
