% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{noise_calc}
\alias{noise_calc}
\title{Environmental noise wrapper}
\usage{
noise_calc(
  resid_time_series,
  predictor,
  noise_method = c("spectrum", "lomb_scargle")
)
}
\arguments{
\item{resid_time_series}{A \code{\link[base]{numeric}} vector containing unpredicted
de-trended residuals as obtained by function \code{\link{monthly_bins}}.}

\item{predictor}{A \code{\link[base]{numeric}} vector containing the time (in days)
for each observation in \code{resid_time_series}. Needs to start at day 0.}

\item{noise_method}{A method for estimating the slope beta. Takes 2 possible
values: \code{'spectrum'} for evenly distributed time series or
\code{'lomb_scargle'} for unevenly distributed ones.}
}
\value{
A \code{\link[base]{list}} containing the beta slope, the original
spec_obj \code{\link[base]{data.frame}}, and the fitted linear model.
}
\description{
Adapts periodogram and calculates the 'colour' of the noise of unpredicted de-trended residuals
}
\details{
This function calculates beta slope (assuming 1/f noise family) on the
residual time series using function \code{\link{spec_slope_get}}.
}
\seealso{
\code{\link{monthly_bins}}, \code{\link{seasonality_and_colour}}, \code{\link{spec_slope_get}}.
}
\author{
Diego Barneche and Scott Burgess.
}
