% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictability.R
\name{predictability}
\alias{predictability}
\title{Calculates environmental predictability components}
\usage{
predictability(
  time_series,
  dates,
  delta,
  is_uneven = FALSE,
  interpolate = FALSE,
  show_warns = TRUE,
  noise_method
)
}
\arguments{
\item{time_series}{A \code{\link[base]{numeric}} vector containing
a raw environmental time series.}

\item{dates}{An vector of class \code{\link[base]{Date}}
of format YYYY-MM-DD (must be in progressive chronological order).}

\item{delta}{Time interval (any unit) of \code{time_series}.}

\item{is_uneven}{Is \code{time_series} even or unevenly distributed in time? Default is \code{\link[base]{FALSE}}.}

\item{interpolate}{Should a linear interpolation be applied to missing values in
\code{time_series}? Irrelevant method if \code{is_uneven} is \code{\link[base]{FALSE}}. Default is \code{\link[base]{FALSE}}.}

\item{show_warns}{Should cautionary warning messages be displayed? Default is \code{\link[base]{TRUE}}.
Strongly recommended for first time users.}

\item{noise_method}{A method for estimating the slope beta. Takes 2 possible
values: \code{'spectrum'} for evenly distributed time series or
\code{'lomb_scargle'} for unevenly distributed ones.}
}
\value{
A \code{\link[base]{data.frame}} with environmental predictability components.
}
\description{
Environmental predictability components
}
\details{
This function currently allows for monthly-based seasonality calculation only.
This algorithm adapts the steps described in \href{http://onlinelibrary.wiley.com/doi/10.1111/ele.12402/abstract}{Marshall and Burgess (2015)} Ecology Letters 18: 1461–0248, doi: \href{http://dx.doi.org/10.1111/ele.12402}{10.1111/ele.12402}.

We advise caution for datasets in which \code{time_series} contains NAs. Particularly,
we strongly advise setting \code{interpolate = FALSE} for \code{time_series} containing large
continuous chunks of NAs because the linear interpolation might introduce substantial bias.

\href{http://onlinelibrary.wiley.com/doi/10.1890/02-3122/abstract}{Vasseur and Yodzis (2004)} Ecology 85: 1146-1152, doi: \href{http://dx.doi.org/10.1890/02-3122}{10.1890/02-3122} recommend time series encompassing at least 128 months or approximately 10 years.

Also note that if \code{dates} starts and ends in different month of the year,
some monthly means are estimated with relatively less precision.
}
\examples{
library(envPred)
data(sst)
# all results return important
# warning messages of which the user should be aware.
# after double-checking that the data falls
# within the recommended specs, set show_warns = FALSE
predictability(sst$time_series,
               sst$dates,
               delta = 1,
               is_uneven = FALSE,
               interpolate = FALSE,
               show_warns = TRUE,
               noise_method = 'spectrum')
predictability(sst$time_series,
               sst$dates,
               delta = 1,
               is_uneven = FALSE,
               interpolate = FALSE,
               show_warns = TRUE,
               noise_method = 'spectrum')
predictability(sst$time_series,
               sst$dates,
               delta = 1,
               is_uneven = FALSE,
               interpolate = FALSE,
               show_warns = TRUE,
               noise_method = 'spectrum')
data(npp)
predictability(npp$time_series,
               npp$dates,
               delta = 8,
               is_uneven = TRUE,
               interpolate = FALSE,
               show_warns = TRUE,
               noise_method = 'lomb_scargle')
predictability(npp$time_series,
               npp$dates,
               delta = 8,
               is_uneven = TRUE,
               interpolate = TRUE,
               show_warns = TRUE,
               noise_method = 'lomb_scargle')
}
\seealso{
\code{\link{seasonality_calc}}, \code{\link{noise_calc}}, \code{\link{colwell74}}.
}
\author{
Diego Barneche and Scott Burgess.
}
