% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colwell.R
\name{predictability_and_colwell}
\alias{predictability_and_colwell}
\title{Calculates environmental predictability components and Colwell (1974) metric}
\usage{
predictability_and_colwell(time_series, dates, n_states, ...)
}
\arguments{
\item{time_series}{A \code{\link[base]{numeric}} vector containing
a raw environmental time series.}

\item{dates}{An vector of class \code{\link[base]{Date}}
of format YYYY-MM-DD (must be in progressive chronological order).}

\item{n_states}{is a \code{\link[base]{numeric}} vector of length 1 containing
a somewhat arbitrary number, as Colwell's method divides a continuous variable
up into discrete states. Default (arbitrary) is 11. See \code{\link{colwell74}}
for Details.}

\item{...}{Additional arguments to \code{\link{predictability}}.}
}
\value{
A \code{\link[base]{data.frame}} with environmental predictability components and
Colwell (1974) metric.
}
\description{
Environmental predictability components
}
\details{
Wrapper function.
}
\seealso{
\code{\link{predictability}},\code{\link{colwell74}}.
}
\author{
Diego Barneche and Scott Burgess.
}
