% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepareLombScargleOutput}
\alias{prepareLombScargleOutput}
\title{Calculates and renames output of Lomb-Scargle Periodogram}
\usage{
prepareLombScargleOutput(residualTimeSeries, ...)
}
\arguments{
\item{residualTimeSeries}{A \code{\link[base]{numeric}} vector containing unpredicted 
detrended residuals as obtained by function \code{\link[envPred]{monthlyBinning}}.}

\item{...}{Additional arguments to \code{\link[lomb]{lsp}}.}
}
\value{
A \code{\link[base]{data.frame}} containing a vector 
of frequencies/periods scanned, and a vector containing the normalised power
corresponding to scanned frequencies/periods.
}
\description{
Adapt Lomb-Scargle Periodogram
}
\seealso{
\code{\link[envPred]{extractSpectrumSlope}}, \code{\link[envPred]{envPredictability}}.
}
\author{
Diego Barneche and Scott Burgess.
}
