% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality_and_colour.R
\name{seasonality_and_colour}
\alias{seasonality_and_colour}
\title{Seasonality and colour of environmental noise}
\usage{
seasonality_and_colour(
  time_series,
  dates,
  delta,
  is_uneven = FALSE,
  interpolate = FALSE,
  show_warns = TRUE,
  noise_method
)
}
\arguments{
\item{time_series}{A \code{\link[base]{numeric}} vector containing
a raw environmental time series.}

\item{dates}{An vector of class \code{\link[base]{Date}}
of format YYYY-MM-DD (must be in progressive chronological order).}

\item{delta}{Time interval (any unit) of \code{time_series}.}

\item{is_uneven}{Is \code{time_series} even or unevenly distributed in time? Default is \code{\link[base]{FALSE}}.}

\item{interpolate}{Should a linear interpolation be applied to missing values in
\code{time_series}? Irrelevant method if \code{is_uneven} is \code{\link[base]{FALSE}}. Default is \code{\link[base]{FALSE}}.}

\item{show_warns}{Should cautionary warning messages be displayed? Default is \code{\link[base]{TRUE}}.
Strongly recommended for first time users.}

\item{noise_method}{A method for estimating the slope theta. Takes 2 possible
values: \code{'spectrum'} for evenly distributed time series or
\code{'lomb_scargle'} for unevenly distributed ones.}
}
\value{
A \code{\link[base]{data.frame}}.
}
\description{
Calculates seasonality and colour of environmental noise.
}
\examples{
library(envPred)
data(sst)
# all results return important
# warning messages of which the user should be aware.
# after double-checking that the data falls
# within the recommended specs, set show_warns = FALSE
fct <- envPred:::seasonality_and_colour
fct(sst$time_series,
    sst$dates,
    delta = 1,
    is_uneven = FALSE,
    interpolate = FALSE,
    show_warns = TRUE,
    noise_method = 'spectrum')
fct(sst$time_series,
    sst$dates,
    delta = 1,
    is_uneven = FALSE,
    interpolate = FALSE,
    show_warns = TRUE,
    noise_method = 'spectrum')
fct(sst$time_series,
    sst$dates,
    delta = 1,
    is_uneven = FALSE,
    interpolate = FALSE,
    show_warns = TRUE,
    noise_method = 'spectrum')
data(npp)
fct(npp$time_series,
    npp$dates,
    delta = 8,
    is_uneven = TRUE,
    interpolate = FALSE,
    show_warns = TRUE,
    noise_method = 'lomb_scargle')
fct(npp$time_series,
    npp$dates,
    delta = 8,
    is_uneven = TRUE,
    interpolate = TRUE,
    show_warns = TRUE,
    noise_method = 'lomb_scargle')
}
\seealso{
\code{\link{env_stats}}, \code{\link{seasonality_calc}}, \code{\link{noise_calc}}.
}
\author{
Diego Barneche and Scott Burgess.
}
