% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{seasonality_calc}
\alias{seasonality_calc}
\title{Calculate the 'colour' of unpredicted de-trended residuals}
\usage{
seasonality_calc(interpolated_seasons, resid_time_series)
}
\arguments{
\item{interpolated_seasons}{A \code{\link[base]{numeric}} vector containing predicted (interpolated)
de-trended residuals as obtained by function \code{\link{monthly_bins}}.}

\item{resid_time_series}{A \code{\link[base]{numeric}} vector containing unpredicted
de-trended residuals as obtained by function \code{\link{monthly_bins}}.}
}
\value{
A \code{\link[base]{list}} containing the sample variance of \code{interpolated_seasons},
\code{resid_time_series}, and the resulting seasonality. See details.
}
\description{
Calculate environmental seasonality
}
\details{
Three types of seasonality are returned currently: an 'absolute' seasonality which corresponds to the sample variance of
\code{interpolated_seasons}; an unbounded seasonality which corresponds to the
ratio between sample variances of \code{interpolated_seasons} and \code{resid_time_series}; and a 'bounded' seasonality which corresponds to the sample variance of \code{interpolated_seasons}
relative to the total summed variances of both \code{interpolated_seasons} and \code{resid_time_series}.
}
\seealso{
\code{\link{monthly_bins}}, \code{\link{predictability}}.
}
\author{
Diego Barneche and Scott Burgess.
}
