% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{spec_output_prep}
\alias{spec_output_prep}
\title{Calculates and renames output of Spectrum Periodogram}
\usage{
spec_output_prep(resid_time_series, ...)
}
\arguments{
\item{resid_time_series}{A \code{\link[base]{numeric}} vector containing unpredicted
de-trended residuals as obtained by function \code{\link{monthly_bins}}.}

\item{...}{Additional arguments to \code{\link[stats]{spectrum}}.}
}
\value{
A \code{\link[base]{data.frame}} containing a vector
of frequencies/periods scanned, and a vector containing the normalised power
corresponding to scanned frequencies/periods.
}
\description{
Adapt Spectrum Periodogram
}
\seealso{
\code{\link{spec_slope_get}}, \code{\link{predictability}}.
}
\author{
Diego Barneche and Scott Burgess.
}
