% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sst}
\alias{sst}
\title{Sea Surface Temperature (SST) Time Series (1997-01-01 - 2007-12-31)}
\format{
A data frame with 4017 rows and 2 variables:
\itemize{
\item time_series: A numeric vector containing a raw environmental time series
\item dates:  A vector of class Date of format YYYY-MM-DD
}
}
\description{
A dataset containing raw time series for SST at one random coordinate.
This time series is evenly distributed with delta time = 1 day.
Data are extracted by Diego Barneche using the noaaErddap GitHub R package.
}
