% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{colwell74}
\alias{colwell74}
\title{Predictability from Colwell (1974)}
\usage{
colwell74(rawTimeSeries, datesVector, nStates = 11)
}
\arguments{
\item{rawTimeSeries}{A \code{\link[base]{numeric}} vector containing 
a raw environmental time series.}

\item{datesVector}{An vector of class \code{\link[base]{Date}} of format YYYY-MM-DD 
(must be in progressive chronological order).}

\item{nStates}{is a \code{\link[base]{numeric}} vector of length 1 containing 
a somewhat arbitrary number, as Colwell's method divides a continuous variable
up into discrete states. Default (arbitrary) is 11. See Details.}
}
\value{
A \code{\link[base]{data.frame}} with three environmental predictability components:
constancy (\code{Colwell_C}), contingency (\code{Colwell_M}) and predictability (\code{Colwell_P}).
}
\description{
Predictability from Colwell (1974)
}
\details{
This algorithm implements the methods described in \href{http://onlinelibrary.wiley.com/doi/10.2307/1940366/abstract}{Colwell (1974)} Ecology 55: 1148–1153, doi: \href{http://dx.doi.org/10.2307/1940366}{10.2307/1940366}.
}
\examples{
library(envPred)
data(sst)
data(npp)
nStates  <-  11
colwell74(sst$rawTimeSeries, sst$datesVector, nStates)
colwell74(npp$rawTimeSeries, npp$datesVector, nStates)
}
\seealso{
\code{\link[envPred]{envPredictability}}.
}
\author{
Diego Barneche and Scott Burgess.
}
