% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{envPredictability}
\alias{envPredictability}
\title{Calculates environmental predictability components}
\usage{
envPredictability(rawTimeSeries, datesVector, delta, isUneven = FALSE,
  interpolate = FALSE, checkPlots = FALSE,
  seasonalityMethod = c("absolute", "unbounded", "bounded"),
  noiseMethod = c("spectrum", "LombScargle"))
}
\arguments{
\item{rawTimeSeries}{A \code{\link[base]{numeric}} vector containing 
a raw environmental time series.}

\item{datesVector}{An object of class \code{\link[base]{Date}} of format YYYY-MM-DD.}

\item{delta}{Time interval (any unit) of \code{rawTimeSeries}.}

\item{isUneven}{Is \code{rawTimeSeries} even or unevenly distributed in time? Default is FALSE.}

\item{interpolate}{Should a linear interpolation be applied to missing values in 
\code{rawTimeSeries}? Irrelevant method if \code{isUneven} is FALSE. Default is FALSE.}

\item{checkPlots}{Should inspection plots be triggered? Default is FALSE. Not recommended 
if applying this task to multiple time series at once.}

\item{seasonalityMethod}{A method for estimating seasonality. Takes 3 possible 
values: \code{'absolute'}, \code{'unbounded'}, or \code{'bounded'} (default).}

\item{noiseMethod}{A method for estimating the slope beta. Takes 2 possible 
values: \code{'spectrum'} for evenly distributed time series or 
\code{'LombScargle'} for unevenly distributed ones.}
}
\value{
A \code{\link[base]{data.frame}} with environmental predictability components.
}
\description{
Environmental predictability components
}
\details{
This function currently allows for monthly seasonality calculation only. 
This algorithm adapts the steps described in \href{http://onlinelibrary.wiley.com/doi/10.1111/ele.12402/abstract}{Marshall and Burgess 2015} Ecology Letters 18: 1461–0248, doi: \href{http://dx.doi.org/10.1111/ele.12402}{10.1111/ele.12402}.
}
\examples{
library(envPred)
data(sst)
envPredictability(sst$rawTimeSeries, sst$datesVector, delta = 1, isUneven = FALSE, interpolate = FALSE, checkPlots = FALSE, seasonalityMethod = 'absolute', noiseMethod = 'spectrum')
envPredictability(sst$rawTimeSeries, sst$datesVector, delta = 1, isUneven = FALSE, interpolate = FALSE, checkPlots = FALSE, seasonalityMethod = 'unbounded', noiseMethod = 'spectrum')
envPredictability(sst$rawTimeSeries, sst$datesVector, delta = 1, isUneven = FALSE, interpolate = FALSE, checkPlots = FALSE, seasonalityMethod = 'bounded', noiseMethod = 'spectrum')
data(npp)
envPredictability(npp$rawTimeSeries, npp$datesVector, delta = 8, isUneven = TRUE, interpolate = FALSE, checkPlots = TRUE, seasonalityMethod = 'unbounded', noiseMethod = 'LombScargle')
envPredictability(npp$rawTimeSeries, npp$datesVector, delta = 8, isUneven = TRUE, interpolate = TRUE, checkPlots = TRUE, seasonalityMethod = 'unbounded', noiseMethod = 'LombScargle')
}
\seealso{
\code{\link[envPred]{envSeasonality}}, \code{\link[envPred]{envNoise}}.
}
\author{
Diego Barneche and Scott Burgess.
}
