% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{envSeasonality}
\alias{envSeasonality}
\title{Calculate the 'color' of unpredicted detrended residuals}
\usage{
envSeasonality(interpolatedSeasons, residualTimeSeries)
}
\arguments{
\item{interpolatedSeasons}{A \code{\link[base]{numeric}} vector containing predicted (interpolated) 
detrended residuals as obtained by function \code{\link[envPred]{monthlyBinning}}.}

\item{residualTimeSeries}{A \code{\link[base]{numeric}} vector containing unpredicted 
detrended residuals as obtained by function \code{\link[envPred]{monthlyBinning}}.}
}
\value{
A \code{\link[base]{list}} containing the sample variance of \code{interpolatedSeasons},
\code{residualTimeSeries}, and the resulting seasonality. See details.
}
\description{
Calculate environmental seasonality
}
\details{
Three types of seasonality are returned currently: an 'absolute' seasonality which corresponds to the sample variance of
\code{interpolatedSeasons}; an unbounded seasonality which corresponds to the 
ratio between sample variances of \code{interpolatedSeasons} and \code{residualTimeSeries}; and a 'bounded' seasonality which corresponds to the sample variance of \code{interpolatedSeasons}
relative to the total summed variances of both \code{interpolatedSeasons} and \code{residualTimeSeries}.
}
\seealso{
\code{\link[envPred]{monthlyBinning}}, \code{\link[envPred]{envPredictability}}.
}
\author{
Diego Barneche and Scott Burgess.
}
