% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extractSpectrumSlope}
\alias{extractSpectrumSlope}
\title{Calculate the 'color' of environmental noise assuming 1/f noise family}
\usage{
extractSpectrumSlope(spectrumObject, plot = FALSE)
}
\arguments{
\item{spectrumObject}{A \code{\link[base]{data.frame}} containing a vector 
of frequencies/periods scanned, and a vector containing the normalised power
corresponding to scanned frequencies/periods.}

\item{plot}{Should inspection plots be triggered? Default is \code{\link[base]{FALSE}}.}
}
\value{
A \code{\link[base]{numeric}} vector of length 1 containing the beta slope.
}
\description{
Adapt Lomb-Scargle Periodogram
}
\details{
Calculates slope of log10(normalised power)~log10(frequencies).
}
\seealso{
\code{\link[envPred]{extractSpectrumSlope}}, \code{\link[envPred]{envPredictability}}.
}
\author{
Diego Barneche and Scott Burgess.
}
