% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_envpred.R
\name{gg_envpred}
\alias{gg_envpred}
\title{Plot predictability object}
\usage{
gg_envpred(object, type = c("detrended", "spectral"))
}
\arguments{
\item{object}{A \code{\link{predictability}} object.}

\item{type}{Either \code{detrended} (default) or \code{spectral}.
See Details.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot predictability object
}
\details{
Two types of plots are provided:
detrended: show the time series after linear
de-trending, including the seasonality interpolation.
spectral: The spectral power density as
a function of frequency.
}
\examples{
\dontrun{
library(envPred)
data(sst)
dat <- predictability(sst$time_series,
                      sst$dates,
                      delta = 1,
                      is_uneven = FALSE,
                      interpolate = FALSE,
                      show_warns = TRUE,
                      noise_method = 'spectrum')

gg_envpred(dat, type = "detrended")
gg_envpred(dat, type = "spectral")
}
}
