% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{spec_slope_get}
\alias{spec_slope_get}
\title{Calculate the 'colour' of environmental noise assuming 1/f noise family}
\usage{
spec_slope_get(spec_obj)
}
\arguments{
\item{spec_obj}{A \code{\link[base]{data.frame}} containing a vector
of frequencies/periods scanned, and a vector containing the normalised power
corresponding to scanned frequencies/periods.}
}
\value{
A \code{\link[base]{numeric}} vector of length 1 containing the beta slope.
}
\description{
Adapt Lomb-Scargle Periodogram
}
\details{
Calculates slope of log10(normalised power)~log10(frequencies).
}
\seealso{
\code{\link{predictability}}.
}
\author{
Diego Barneche and Scott Burgess.
}
